+++ src/posting/widgets/websocket/websocket_container.py
@@ -12,8 +12,8 @@ class WebSocketContainer(Vertical):
         yield Replies(classes="section")
 
     @on(Replies.Incoming)
-    def on_incoming(self, event: Replies.Incoming) -> None:
-        self.replies.add_reply(event)
+    async def on_incoming(self, event: Replies.Incoming) -> None:
+        await self.replies.add_reply(event)
 
     @property
     def composer(self) -> WebsocketComposer: